pyproject.toml
[project]
name = "ai-coding-agent-glassbox"
version = "0.1.0"
description = "AI Coding Agent - Glass Box Implementation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",  # For Llama (Ollama)
    "python-dotenv>=1.0.0",  # For .env file support
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ai-agent = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
